+++ R/rtf_colheader.R
@@ -18,7 +18,8 @@
 #' @title Add Column Header Attributes to Table
 #'
 #' @param tbl A data frame.
-#' @param colheader A character string that uses " | " to separate column names. Default is NULL for a blank column header.
+#' @param colheader A character string that uses " | " to separate column names.
+#' Default is NULL for a blank column header.
 #' @param col_rel_width A Column relative width in a vector e.g. c(2,1,1) refers to 2:1:1.
 #'                      Default is NULL for equal column width.
 #' @inheritParams rtf_footnote
@@ -26,7 +27,8 @@
 #' @section Specification:
 #' \if{latex}{
 #'  \itemize{
-#'    \item Input checks using \code{check_args()}, \code{match_arg()} and \code{stopifnot()}. The required argument is \code{tbl}, i.e. A data frame must define by \code{tbl}.
+#'    \item Input checks using \code{check_args()}, \code{match_arg()} and \code{stopifnot()}.
+#'    The required argument is \code{tbl}, i.e. A data frame must define by \code{tbl}.
 #'    \item Set default page attributes and register use_color attribute.
 #'    \item Define column header attributes of \code{tbl} based on the input.
 #'    \item Return \code{tbl}.
@@ -48,42 +50,33 @@
 #'   attr("rtf_colheader")
 #' @export
 rtf_colheader <- function(tbl,
-
                           colheader = NULL,
                           col_rel_width = NULL,
-
                           border_left = "single",
                           border_right = "single",
                           border_top = "single",
                           border_bottom = "",
-
                           border_color_left = NULL,
                           border_color_right = NULL,
                           border_color_top = NULL,
                           border_color_bottom = NULL,
-
                           border_width = 15,
-
                           cell_height = 0.15,
                           cell_justification = "c",
+                          cell_vertical_justification = "bottom",
                           cell_nrow = NULL,
-
                           text_font = 1,
                           text_format = NULL,
                           text_font_size = 9,
-
                           text_color = NULL,
                           text_background_color = NULL,
                           text_justification = "c",
-
                           text_indent_first = 0,
                           text_indent_left = 0,
                           text_indent_right = 0,
-
                           text_space = 1,
                           text_space_before = 15,
                           text_space_after = 15,
-
                           text_convert = TRUE) {
 
   # Check argument type
@@ -91,6 +84,9 @@ rtf_colheader <- function(tbl,
 
   check_args(colheader, type = c("character"))
   check_args(col_rel_width, type = c("integer", "numeric"))
+
+  # Convert tbl to a data frame, each column is a character
+  if (any(class(tbl) %in% "data.frame")) tbl <- as.data.frame(tbl, stringsAsFactors = FALSE)
 
   # Set Default Page Attributes
   if (is.null(attr(tbl, "page"))) {
@@ -110,51 +106,43 @@ rtf_colheader <- function(tbl,
 
   # Define text attributes
   colheader <- obj_rtf_text(colheader,
-
     text_font,
     text_format,
     text_font_size,
     text_color,
     text_background_color,
     text_justification,
-
     text_indent_first,
     text_indent_left,
     text_indent_right,
     text_space = 1,
     text_space_before,
     text_space_after,
-
     text_new_page = FALSE,
     text_hyphenation = TRUE,
-
     text_convert = text_convert
   )
   if (attr(colheader, "use_color")) attr(tbl, "page")$use_color <- TRUE
 
   # Define border attributes
   colheader <- obj_rtf_border(colheader,
-
     border_left,
     border_right,
     border_top,
     border_bottom,
     border_first = NULL,
     border_last  = NULL,
-
     border_color_left,
     border_color_right,
     border_color_top,
     border_color_bottom,
-
     border_color_first = NULL,
     border_color_last  = NULL,
-
-    border_width,
-
-    cell_height,
-    cell_justification,
-    cell_nrow
+    border_width = border_width,
+    cell_height = cell_height,
+    cell_justification = cell_justification,
+    cell_vertical_justification = cell_vertical_justification,
+    cell_nrow = cell_nrow
   )
 
   if (attr(colheader, "use_color")) attr(tbl, "page")$use_color <- TRUE