+++ R/rtf_footnote.R
@@ -53,10 +53,12 @@
 #'                            equal to number of columns displayed e.g. c("white","red","blue").
 #'                            All possible input can be found in `grDevices::colors()`.
 #' @param border_width Border width in twips. Default is 15 for 0.0104 inch.
-#' @param cell_justification Justification type for cell. Default is "c" for center justification.
+#' @param cell_justification Justification type for cell.
 #'                           All possible input can be found in `r2rtf:::justification()$type`.
+#' @param cell_vertical_justification Vertical justification type for cell.
+#'                           All possible input can be found in `r2rtf:::vertical_justification()$type`.
 #' @param cell_height Cell height in inches. Default is 0.15 for 0.15 inch.
-#' @param cell_nrow Number of rows required in each cell. Default is NULL.
+#' @param cell_nrow Number of rows required in each cell.
 #' @param text_justification Justification type for text. Default is "c" for center justification.
 #'                           To vary text justification by column, use character vector with
 #'                           length of vector equal to number of columns displayed e.g. c("c","l","r").
@@ -65,7 +67,7 @@
 #'                  by column, use numeric vector with length of vector equal to number of
 #'                  columns displayed e.g. c(1,2,3).All possible input can be found
 #'                  in `r2rtf:::font_type()$type`.
-#' @param text_font_size Text font size. Default is 9. To vary text font size by column, use
+#' @param text_font_size Text font size.  To vary text font size by column, use
 #'                       numeric vector with length of vector equal to number of columns
 #'                       displayed e.g. c(9,20,40).
 #' @param text_format Text format type. Default is NULL for normal. Combination of format type
@@ -86,10 +88,10 @@
 #' @param text_indent_right A value of text right indent.
 #' @param text_indent_reference The reference start point of text indent. Accept `table` or `page_margin`
 #' @param text_space Line space between paragraph in twips. Default is 0.
-#' @param text_space_before Line space before a paragraph in twips. Default is 15.
-#' @param text_space_after Line space after a paragraph in twips. Default is 15.
+#' @param text_space_before Line space before a paragraph in twips.
+#' @param text_space_after Line space after a paragraph in twips.
 #' @param as_table A logical value to display it as a table.
-#' @param text_convert A logical value to convert special characters. Default is TRUE.
+#' @param text_convert A logical value to convert special characters.
 #'
 #' @section Specification:
 #' \if{latex}{
@@ -110,55 +112,49 @@
 #'   attr("rtf_footnote")
 #' @export
 rtf_footnote <- function(tbl,
-
                          footnote = "",
-
                          border_left = "single",
                          border_right = "single",
                          border_top = "",
                          border_bottom = "single",
-
                          border_color_left = NULL,
                          border_color_right = NULL,
                          border_color_top = NULL,
                          border_color_bottom = NULL,
-
                          border_width = 15,
-
                          cell_height = 0.15,
                          cell_justification = "c",
+                         cell_vertical_justification = "top",
                          cell_nrow = NULL,
-
                          text_font = 1,
                          text_format = NULL,
                          text_font_size = 9,
                          text_color = NULL,
                          text_background_color = NULL,
                          text_justification = "l",
-
                          text_indent_first = 0,
                          text_indent_left = 0,
                          text_indent_right = 0,
                          text_indent_reference = "table",
-
                          text_space = 1,
                          text_space_before = 15,
                          text_space_after = 15,
-
                          text_convert = TRUE,
-
                          as_table = TRUE) {
 
   # Check argument type
   check_args(footnote, type = "character")
   check_args(as_table, type = "logical")
 
+  # Convert tbl to a data frame, each column is a character
+  if (any(class(tbl) %in% "data.frame")) tbl <- as.data.frame(tbl, stringsAsFactors = FALSE)
+
   # Define proper justification reference
-  if (text_justification == "l" & (! as_table) ) {
+  if (text_justification == "l" & (!as_table)) {
     text_indent_left <- text_indent_left + footnote_source_space(tbl, text_indent_reference)
   }
 
-  if (text_justification == "r" & (! as_table) ) {
+  if (text_justification == "r" & (!as_table)) {
     text_indent_right <- text_indent_right + footnote_source_space(tbl, text_indent_reference)
   }
 
@@ -169,14 +165,12 @@ rtf_footnote <- function(tbl,
 
   # Define text object
   footnote <- obj_rtf_text(footnote,
-
     text_font,
     text_format,
     text_font_size,
     text_color,
     text_background_color,
     text_justification,
-
     text_indent_first,
     text_indent_left,
     text_indent_right,
@@ -185,7 +179,6 @@ rtf_footnote <- function(tbl,
     text_space_after,
     text_new_page = FALSE,
     text_hyphenation = TRUE,
-
     text_convert = text_convert
   )
 
@@ -194,27 +187,23 @@ rtf_footnote <- function(tbl,
   # Define border object
   if (as_table) {
     footnote <- obj_rtf_border(footnote,
-
       border_left,
       border_right,
       border_top,
       border_bottom,
       border_first = NULL,
       border_last  = NULL,
-
       border_color_left,
       border_color_right,
       border_color_top,
       border_color_bottom,
-
       border_color_first = NULL,
       border_color_last  = NULL,
-
-      border_width,
-
-      cell_height,
-      cell_justification,
-      cell_nrow
+      border_width = border_width,
+      cell_height = cell_height,
+      cell_justification = cell_justification,
+      cell_vertical_justification = cell_vertical_justification,
+      cell_nrow = cell_nrow
     )
 
     if (attr(footnote, "use_color")) attr(tbl, "page")$use_color <- TRUE