+++ inst/doc/r2rtf-cran.R
@@ -1,6 +1,9 @@
 ## ---- eval = FALSE------------------------------------------------------------
-#  head(tbl) %>%
-#    rtf_body() %>%                     # Step 1 Add table  attributes
-#    rtf_encode() %>%                   # Step 2 Convert attributes to RTF encode
-#    write_rtf("rtf/intro-ae1.rtf")     # Step 3 Write to a .rtf file
+#  library(dplyr)
+#  library(r2rtf)
+#  
+#  head(iris) %>%
+#    rtf_body() %>% # Step 1 Add attributes
+#    rtf_encode() %>% # Step 2 Convert attributes to RTF encode
+#    write_rtf(file = "ex-tbl.rtf") # Step 3 Write to a .rtf file